# perf improvements for depthwise convolutions