--- a/torch/csrc/autograd/functions/convolution.cpp
+++ b/torch/csrc/autograd/functions/convolution.cpp
@@ -365,10 +365,8 @@ auto ConvForward::apply(const variable_list& inputs) -> variable_list {
 // For Convolution strategies that don't implicitly handle grad_bias, we add a helper
 // function here to perform it using simple Tensor operators
 static at::Tensor compute_grad_bias(const at::Tensor& grad_output) {
-  // grad_output is in N, C, H, W, we re-shape and make contiguous
-  at::Tensor transposed = grad_output.transpose(0, 1).contiguous();
-  // sum across all of the channels and add to grad_bias
-  return transposed.view({transposed.size(0), -1}).sum(1);
+  // grad_output is in N, C, H, W, we re-shape and reduce over spatial dims and batches 
+  return grad_output.contiguous().view({grad_output.size(0), grad_output.size(1), -1}).sum(0).sum(1);
 }
 
 // ConvBackward implementation