@elliothe 2017-11-05
In order to use this new depthwise conv, do I have to install the cudnn v7?