@soumith 2017-11-06
@elliothe no.