@elliothe 2017-11-06
@soumith  in order to get the fastest depthwise convolution in pytorch, what should I do now? Since I have installed the lastest pytorch from source, with CUDA 8.0 and cudnn v6. However, if I try to replace the normal spatial conv through combining depthwise conv and summation, the simulation speed is more than tens times slower. Is this speed degradation normal?