@soumith 2017-11-06
@elliothe how did you actually "replace" normal spatial conv with depthwise conv?

All you need to do to use `depthwise conv` is by using the `groups` parameter of `nn.Conv2d`