@elliothe 2017-11-06
@soumith Normal spatial convolution is `nn.Conv2d(input_channels,output_channels)`, where I rewirte it as:
```
nn.Conv2d(input_channels, input_channels*output_channels, groups = input_channels)
nn.ReLU()   # wanna test the intermediate activation function
nn.view(-1, input_channels, output_channels, Hin, Win).sum(1)
```
When the input_channels and output_channels are large, the speed degrades a lot.