@csarofeen 2017-11-06
Wouldn't you also see this "degradation" with nn.Conv2d(input_channels, input_channels*output_channels)?