@ngimel 2017-11-06
That's a fair point, though, currently when deciding between cudnn and depthwise-separable kernels the only thing that's checked is that number_of_groups = input_channels, however, when depth_multiplier is large (that is, number of output channels = <large numer> * number of input channels), using cudnn and just dispatching a number of regular convolution kernels might be faster. Are layers like this encountered in practice?