@elliothe 2017-11-06
@ngimel I believe this is going to be a research direction, since the depthwise-separable kernels become popular due to some recent the-state-of-art NNs, like mobilenets, xception. It would be very helpful for researchers, if pytorch can support such kind of function efficiently.