@ngimel 2017-11-06
In Mobilenet and xception the number of output channels is equal to the number of input channels (may be in some xception layers it's 2*input_channels, I don't remember off the top of my head). For those layers the kernels in pytorch provide much better performance than what used to be available (the benchmarks for mobilenet comparing with the previous cudnn implementation are in #3057, this PR slightly improves on #3057).